# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/weekly-plans/generate": {
      "post": {
        "description": "Placeholder — returns 501. Auto-generation strategy is undecided. Use `POST /weekly-plans` to create plans manually.",
        "operationId": "WeeklyPlansController_generate",
        "parameters": [],
        "responses": {
          "501": {
            "description": "Not implemented."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Auto-generate weekly plans [NOT IMPLEMENTED]",
        "tags": [
          "Weekly Plans"
        ]
      }
    },
    "/weekly-plans": {
      "post": {
        "description": "Creates a draft weekly plan with one or more items. Caller must be principal, VP, supervisor in scope, or any teacher assigned to the halaqa. Returns 409 if a plan already exists for the same student/halaqa/week.",
        "operationId": "WeeklyPlansController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWeeklyPlanDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeeklyPlanDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation error or invalid verse range."
          },
          "403": {
            "description": "No halaqa scope."
          },
          "409": {
            "description": "Plan already exists. Response includes existing_plan_id."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Create a weekly plan",
        "tags": [
          "Weekly Plans"
        ]
      },
      "get": {
        "description": "Returns a paginated list of weekly plans visible to the caller.\n- **principal / vice_principal**: all plans in the school.\n- **supervisor**: plans for halaqat they supervise.\n- **teacher**: plans for halaqat they are assigned to.\n- **parent**: plans for their linked students only.",
        "operationId": "WeeklyPlansController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Defaults to 1.",
            "schema": {
              "minimum": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Defaults to 20, capped at 100.",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "student_id",
            "required": false,
            "in": "query",
            "description": "Filter by student.",
            "schema": {
              "example": 42,
              "type": "number"
            }
          },
          {
            "name": "halaqa_id",
            "required": false,
            "in": "query",
            "description": "Filter by halaqa.",
            "schema": {
              "example": 3,
              "type": "number"
            }
          },
          {
            "name": "week_start_date",
            "required": false,
            "in": "query",
            "description": "Filter by week start date.",
            "schema": {
              "format": "date",
              "example": "2026-05-09",
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "example": "draft",
              "type": "string",
              "enum": [
                "draft",
                "approved"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeeklyPlanListData"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List weekly plans (role-scoped)",
        "tags": [
          "Weekly Plans"
        ]
      }
    },
    "/weekly-plans/{id}": {
      "get": {
        "description": "Returns the plan with all its items. Out-of-scope or cross-school access returns 404 (never 403).",
        "operationId": "WeeklyPlansController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Weekly plan ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeeklyPlanDto"
                }
              }
            }
          },
          "404": {
            "description": "Not found or out of scope."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get weekly plan detail",
        "tags": [
          "Weekly Plans"
        ]
      },
      "delete": {
        "description": "Permanently deletes the plan and its items. Caller must be principal, VP, supervisor in scope, or any teacher assigned to the halaqa. This is a hard delete — the plan and its items are removed from the database and cannot be restored.",
        "operationId": "WeeklyPlansController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Weekly plan ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiMessage"
                }
              }
            }
          },
          "403": {
            "description": "No halaqa scope."
          },
          "404": {
            "description": "Not found."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Delete (permanently) a weekly plan",
        "tags": [
          "Weekly Plans"
        ]
      }
    },
    "/weekly-plans/{id}/approve": {
      "post": {
        "description": "Locks the plan structure and runs reconciliation for every item. Caller must be principal, VP, supervisor in scope, or any teacher assigned to the halaqa. Returns 400 if already approved.",
        "operationId": "WeeklyPlansController_approve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Weekly plan ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeeklyPlanDto"
                }
              }
            }
          },
          "400": {
            "description": "Already approved."
          },
          "403": {
            "description": "No halaqa scope."
          },
          "404": {
            "description": "Not found."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Approve a weekly plan",
        "tags": [
          "Weekly Plans"
        ]
      }
    },
    "/weekly-plans/{id}/unapprove": {
      "post": {
        "description": "Reverts the plan to draft status. Caller must be principal, VP, supervisor in scope, or any teacher assigned to the halaqa. `approved_by` is preserved for audit purposes. Returns 400 if not currently approved.",
        "operationId": "WeeklyPlansController_unapprove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Weekly plan ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeeklyPlanDto"
                }
              }
            }
          },
          "400": {
            "description": "Not currently approved."
          },
          "403": {
            "description": "No halaqa scope."
          },
          "404": {
            "description": "Not found."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Unapprove a weekly plan",
        "tags": [
          "Weekly Plans"
        ]
      }
    },
...
  },
  "info": {
    "title": "Halaqa Backend API",
    "description": "Authentication, user management, and school operations for the Halaqa system. All JSON responses are wrapped in a uniform envelope: `{ \"code\": <http>, \"data\": ... }` on success, `{ \"code\": <http>, \"message\": \"...\" }` on errors. Refresh tokens travel as HttpOnly cookies on `/auth/*`; access tokens use Bearer auth.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refresh_token"
      }
    },
    "schemas": {
...
      "ApiMessage": {
        "type": "object",
        "properties": {}
      },
...
      "CreateWeeklyPlanItemDto": {
        "type": "object",
        "properties": {
          "track_type": {
            "type": "string",
            "enum": [
              "Hifz",
              "Near",
              "Far"
            ],
            "example": "Hifz"
          },
          "day_of_week": {
            "type": "number",
            "example": 2,
            "minimum": 0,
            "maximum": 6,
            "description": "0=Saturday … 6=Friday."
          },
          "order": {
            "type": "number",
            "example": 0,
            "minimum": 0,
            "default": 0,
            "description": "Reconciliation priority within the same day + track (lower = first). Defaults to 0."
          },
          "start_surah": {
            "type": "number",
            "example": 1,
            "minimum": 1,
            "maximum": 114
          },
          "start_verse": {
            "type": "number",
            "example": 1,
            "minimum": 1
          },
          "end_surah": {
            "type": "number",
            "example": 1,
            "minimum": 1,
            "maximum": 114,
            "description": "Must be >= start_surah."
          },
          "end_verse": {
            "type": "number",
            "example": 7,
            "minimum": 1
          }
        },
        "required": [
          "track_type",
          "day_of_week",
          "start_surah",
          "start_verse",
          "end_surah",
          "end_verse"
        ]
      },
      "CreateWeeklyPlanDto": {
        "type": "object",
        "properties": {
          "student_id": {
            "type": "number",
            "example": 42,
            "description": "Student ID. Must belong to the caller's school."
          },
          "halaqa_id": {
            "type": "number",
            "example": 3,
            "description": "Halaqa ID. Caller must have scope on this halaqa."
          },
          "week_start_date": {
            "type": "string",
            "format": "date",
            "example": "2026-05-09",
            "description": "Saturday that starts the week."
          },
          "items": {
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateWeeklyPlanItemDto"
            }
          }
        },
        "required": [
          "student_id",
          "halaqa_id",
          "week_start_date",
          "items"
        ]
      },
      "WeeklyPlanItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "track_type": {
            "type": "string",
            "enum": [
              "Hifz",
              "Near",
              "Far"
            ],
            "example": "Hifz"
          },
          "day_of_week": {
            "type": "number",
            "example": 2,
            "description": "0=Saturday … 6=Friday."
          },
          "order": {
            "type": "number",
            "example": 0,
            "description": "Reconciliation priority within the same day + track (lower = first)."
          },
          "start_surah": {
            "type": "number",
            "example": 1
          },
          "start_verse": {
            "type": "number",
            "example": 1
          },
          "end_surah": {
            "type": "number",
            "example": 1
          },
          "end_verse": {
            "type": "number",
            "example": 7
          },
          "total_verses": {
            "type": "number",
            "example": 7
          },
          "achieved_verses": {
            "type": "number",
            "example": 4
          },
          "status": {
            "type": "string",
            "enum": [
              "due",
              "overdue",
              "partial",
              "completed"
            ],
            "example": "partial"
          },
          "is_manual_override": {
            "type": "boolean",
            "example": false,
            "description": "True if range was edited after item creation."
          },
          "created_at": {
            "type": "string",
            "example": "2026-05-11T10:00:00.000Z"
          }
        },
        "required": [
          "id",
          "track_type",
          "day_of_week",
          "order",
          "start_surah",
          "start_verse",
          "end_surah",
          "end_verse",
          "total_verses",
          "achieved_verses",
          "status",
          "is_manual_override",
          "created_at"
        ]
      },
      "WeeklyPlanDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "student_id": {
            "type": "number",
            "example": 42
          },
          "halaqa_id": {
            "type": "number",
            "example": 3
          },
          "week_start_date": {
            "type": "string",
            "format": "date",
            "example": "2026-05-09"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "approved"
            ],
            "example": "draft"
          },
          "approved_by": {
            "type": "object",
            "nullable": true,
            "example": 7
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WeeklyPlanItemDto"
            }
          },
          "created_at": {
            "type": "string",
            "example": "2026-05-09T10:00:00.000Z"
          }
        },
        "required": [
          "id",
          "student_id",
          "halaqa_id",
          "week_start_date",
          "status",
          "approved_by",
          "items",
          "created_at"
        ]
      },
      "WeeklyPlanListData": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WeeklyPlanDto"
            }
          },
          "total": {
            "type": "number",
            "example": 10
          },
          "page": {
            "type": "number",
            "example": 1
          },
          "limit": {
            "type": "number",
            "example": 20
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "limit"
        ]
      },
...
    }
  }
}